Move MapContainer tests into RouterTest
